use booleans instead of numbers for matching
the -1 comparison is due to using indexOf. Instead, let's compare to -1 at the site of indexOf to get a boolean, and also use a boolean in the other branches of the if statement. this way we can simplify the final boolean expression, and set us up for a refactor in the next commit by extracting a contains() function