collapse the if/else statements to an equivalent boolean expression
the initial branch of the if statement already says if this OR that -> true, so that's an easy one. the else branch of the statement behaves exactly like an OR (think how it gets executed, if the if statement is true then it evaluates the else: false || true). so we can merge that in too. aside from the reasons for the change, i think the result is also easier to read/understand (it reads like you would explain the logic to someone), and again removes the need to have explanatory comments.